requirements.txt
boto3==1.34.155
botocore==1.34.155
grpcio==1.65.4
grpcio-tools==1.65.4
jmespath==1.0.1
protobuf==5.27.3
pyaml==24.7.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.1
redis==5.0.8
s3transfer==0.10.2
setuptools==72.1.0
six==1.16.0
urllib3==2.2.2
